+++ app/api/auth/verify-otp/route.js
@@ -19,46 +19,47 @@ export async function POST(request) {
             return response(false, 401, 'Invalid or missing input field.', validatedData.error)
         }
 
-const { email, otp } = validatedData.data
+        const { email, otp } = validatedData.data
 
-const getOtpData = await OTPModel.findOne({ email, otp })
-if (!getOtpData) {
-    return response(false, 404, 'Invalid or expired otp.')
-}
+        const getOtpData = await OTPModel.findOne({ email, otp })
+        if (!getOtpData || new Date() > new Date(getOtpData.expiresAt)) {
+            if (getOtpData) await getOtpData.deleteOne()
+            return response(false, 404, 'Invalid or expired otp.')
+        }
 
-const getUser = await UserModel.findOne({ deletedAt: null, email }).lean()
-if (!getUser) {
-    return response(false, 404, 'User not found.')
-}
-const loggedInUserData = {
-    _id: getUser._id,
-    role: getUser.role,
-    name: getUser.name,
-    avatar: getUser.avatar,
-}
+        const getUser = await UserModel.findOne({ deletedAt: null, email }).lean()
+        if (!getUser) {
+            return response(false, 404, 'User not found.')
+        }
+        const loggedInUserData = {
+            _id: getUser._id,
+            role: getUser.role,
+            name: getUser.name,
+            avatar: getUser.avatar,
+        }
 
 
-const secret = new TextEncoder().encode(process.env.SECRET_KEY)
-const token = await new SignJWT(loggedInUserData)
-.setIssuedAt()
-.setExpirationTime('24h')
-.setProtectedHeader({alg: 'HS256'})
-.sign(secret)
+        const secret = new TextEncoder().encode(process.env.SECRET_KEY)
+        const token = await new SignJWT(loggedInUserData)
+            .setIssuedAt()
+            .setExpirationTime('24h')
+            .setProtectedHeader({ alg: 'HS256' })
+            .sign(secret)
 
-const cookieStore = await cookies()
+        const cookieStore = await cookies()
 
-cookieStore.set({
-    name: 'access_token',
-    value: token,
-    httpOnly: process.env.NODE_ENV === 'production',
-    path: '/',
-    secure: process.env.NODE_ENV === 'production',
-    sameSite: 'lax',
-})
-// remove otp after validation
-await getOtpData.deleteOne()
+        cookieStore.set({
+            name: 'access_token',
+            value: token,
+            httpOnly: process.env.NODE_ENV === 'production',
+            path: '/',
+            secure: process.env.NODE_ENV === 'production',
+            sameSite: 'lax',
+        })
+        // remove otp after validation
+        await getOtpData.deleteOne()
 
-return response(true, 200, 'Login successfull.', loggedInUserData)
+        return response(true, 200, 'Login successfull.', loggedInUserData)
 
     } catch (error) {
         return catchError(error)
